Abstract types from Engine
